Plans: rename lib from currency-formatter to format-currency
