+++ client/lib/format-currency/README.md
@@ -1,21 +1,21 @@
-Currency Formatter
+Format Currency
 ==========
 Given a currency code, this library will take in a number and display it as money correctly.
 
 Usage
 ==========
 ```javascript
-import currencyFormatter from 'lib/currency-formatter';
+import formatCurrency from 'lib/format-currency';
 
-const USD = currencyFormatter( 9800900.32, 'USD' ); // '$9,800,900.32'
-const EUR = currencyFormatter( 9800900.32, 'EUR' ); // '€9.800.900,32'
-const JPY = currencyFormatter( 9800900.32, 'JPY' ); // '¥9,800,900'
+const USD = formatCurrency( 9800900.32, 'USD' ); // '$9,800,900.32'
+const EUR = formatCurrency( 9800900.32, 'EUR' ); // '€9.800.900,32'
+const JPY = formatCurrency( 9800900.32, 'JPY' ); // '¥9,800,900'
 ```
 
 Or
 
 ```javascript
-import { getCurrencyObject } from 'lib/currency-formatter';
+import { getCurrencyObject } from 'lib/format-currency';
 const USD = getCurrencyObject( 9800900.32, 'USD' ); // { symbol: '$', integer: '9,800,900', fraction: '.32', sign: '' }
 
 ```
