+++ client/lib/format-currency/index.js
